move methods to fetch CPT and taxonomies in a trait
